+++ egui/src/containers/window.rs
@@ -882,14 +882,10 @@ impl TitleBar {
 
         if let Some(content_response) = &content_response {
             // paint separator between title and content:
-            let left = outer_rect.left();
-            let right = outer_rect.right();
             let y = content_response.rect.top() + ui.spacing().item_spacing.y * 0.5;
             // let y = lerp(self.rect.bottom()..=content_response.rect.top(), 0.5);
-            ui.painter().line_segment(
-                [pos2(left, y), pos2(right, y)],
-                ui.visuals().widgets.noninteractive.bg_stroke,
-            );
+            let stroke = ui.visuals().widgets.noninteractive.bg_stroke;
+            ui.painter().hline(outer_rect.x_range(), y, stroke);
         }
 
         if ui